ListDisplay Item Displaying
The ListDisplay now supports displaying items.

There's still a lot to do for this class, but it now has the most basic functionality of presenting data.

The current build is a proof of concept where the view shifts with the big bottom button and the buttons of each row self update the contents of the same row. Where, all definitions for logic are handled almost entirely with minimal code in the TestClass.cs and a small amount in the TestClassDisplay to tell the ListDisplay what data it's supposed to be displaying.