ID System
I should have made smaller commits but I didn't have time; was busy working. And this project is just kinda silly and for fun anyway so I didn't bother.

Anywho, I added an ID system. It's very complicated and I wholeheartedly expect it to break when I dont want it to. But oh well, it seems to work for now. This will be useful in the future when I need to start associating data with other data.

I've attached a primary key to the player so the player can later be referenced.