Property Sections
Added a whole bunch of stuff to do with Property Sections and a whole bunch of small changes around in different places to do with property sections.

They're not completely done, but they are virtually there. They are completely functional (albeit a little buggy with their error detection, but it works in the end as long as everything is right). The only thing I'll be adding is the ability to move sections up and down in the list (probably unless I dont want to).

1 step at a time to finishing the property additions screen.