Property Manager *Completed
The Property Manager is now *completed. I'm sure there'll be further changes down the line when I work to implement property viewing (namely, creating a property manager with with preset values), but that's not needed yet.

The next part of the property manager will be done after the Property class works. They need to implement each other so that the manager can successfully create a Property which can be saved and displayed in a list.