Property Modification Switch Management
So, so many changes.

Modified parts of most switch management related systems to make ListBrowser and specifically the PropertyManager inside of it in this case compatible with switch management.

Contents of the ListBrowser will be forwarded through the switch management system where, if changes are made to the content the listbrowser will use the switch management implementation to get confirmation of switching to a different list element view (display content) before switching.