Basic Land Testing
Added basic functionality for the ILandArea interface and created Land to test it.

Modified the Coordinates system to include a "Bounds" based system of cubes or square sections.

Split ICoordinate into IFlatCoordinate and ICoordinate.

Added a universal governmental ID marker at 10p...1 and the ability to hide this ID from the playerslist display.