Construct Highest Identifier
Created a function which attempts to fully (re)construct an identifier from and ID based on the available data.

This was an annoying point for me, where trying to display which ID is having an error causes the ID to error because the ID has an error. This function circumvents that issue by just checking for errors and display invalid messages for the portion of ID which cannot be found or otherwise has errors.